% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_paper.R
\name{compile_paper}
\alias{compile_paper}
\title{Compile the `BundibugyoNPI2026.tex` \LaTeX document into a PDF}
\usage{
compile_paper()
}
\value{
Nothing, called for side-effects.
}
\description{
Compile the `BundibugyoNPI2026.tex` \LaTeX document into a PDF
}
\keyword{internal}
