% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_sim_cond.R
\name{scenario_sim_cond}
\alias{scenario_sim_cond}
\title{[ringbp::scenario_sim()] with conditioning on each outbreak}
\usage{
scenario_sim_cond(
  n,
  initial_cases,
  offspring,
  delays,
  event_probs,
  interventions,
  sim,
  cond
)
}
\arguments{
\item{n}{a positive \code{integer} scalar: number of simulations to run}

\item{initial_cases}{a non-negative \code{integer} scalar: number of initial
or starting cases which are all assumed to be missed by contact tracing
(i.e. tracing ascertainment = 0).}

\item{offspring}{a \code{list} with class \verb{<ringbp_offspring_opts>}: the
offspring distribution \code{function}s for the \pkg{ringbp} model, returned
by \code{\link[ringbp:offspring_opts]{offspring_opts()}}. Contains three elements: \code{community}, \code{isolated},
and \code{asymptomatic}}

\item{delays}{a \code{list} with class \verb{<ringbp_delay_opts>}: the
delay distribution \code{function}s for the \pkg{ringbp} model, returned
by \code{\link[ringbp:delay_opts]{delay_opts()}}. Contains 4 elements: \code{incubation_period},
\code{onset_to_isolation}, \code{latent_period} and \code{onset_to_self_isolation}}

\item{event_probs}{a \code{list} with class \verb{<ringbp_event_prob_opts>}: the
event probabilities for the \pkg{ringbp} model, returned by
\code{\link[ringbp:event_prob_opts]{event_prob_opts()}}. Contains 5 elements: \code{asymptomatic},
\code{presymptomatic_transmission}, \code{alpha}, \code{symptomatic_traced} and
\code{symptomatic_self_isolate}}

\item{interventions}{a \code{list} with class \verb{<ringbp_intervention_opts>}:
the intervention settings for the \pkg{ringbp} model, returned by
\code{\link[ringbp:intervention_opts]{intervention_opts()}}. Contains 2 elements: \code{quarantine} and
\code{test_sensitivity}}

\item{sim}{a \code{list} with class \verb{<ringbp_sim_opts>}: the simulation control
options for the \pkg{ringbp} model, returned by \code{\link[ringbp:sim_opts]{sim_opts()}}}

\item{cond}{A list including:
* `max_attempts`: the maximum number of attempts to simulate each outbreak
that meets conditioning before erroring
* `target_weeks`: an `integer` vector of weeks during which the cumulative
outbreak size is checked against `target_cases`
* `target_cases`: a `numeric` vector of length 2 with the lower and upper
bounds (inclusive) of the cumulative outbreak size for the conditioning}
}
\value{
A `data.table`
}
\description{
[ringbp::scenario_sim()] with conditioning on each outbreak
}
